태성책임님, 유리님 리뷰 수정 사항 적용

+++ RaywattApp/RaywattApp/Common/Bases/Constants.cs
@@ -416,9 +416,9 @@ namespace RaywattApp.Common.Bases
         public const int commandPacketSize = 5;
 
         //AngioCoRegistration
-        public const int ellipsePathWidth = 2;
-        public const int ellipsePathHeight = 2;
-        public const int ellipseTrackWidth = 6;
-        public const int ellipseTrackHeight = 6;
+        public const double ellipsePathWidth = 2;
+        public const double ellipsePathHeight = 2;
+        public const double ellipseTrackWidth = 6;
+        public const double ellipseTrackHeight = 6;
     }
 }
